+++ common/src/main/java/com/st0x0ef/stellaris/mixin/CandleBlockMixin.java
@@ -2,7 +2,6 @@ package com.st0x0ef.stellaris.mixin;
 
 import com.st0x0ef.stellaris.common.utils.PlanetUtil;
 import net.minecraft.core.BlockPos;
-import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.ItemInteractionResult;
 import net.minecraft.world.entity.player.Player;